+++ packages/components/src/components/Card/index.ts
@@ -1,2 +1,8 @@
 export { Card }           from './Card'
-export type { CardProps, CardVariant, CardElement } from './Card'
+export type {
+  CardProps,
+  CardOwnProps,
+  CardAsChildProps,
+  CardVariant,
+  CardElement,
+} from './Card'